feat: add tx details for migration
