+++ apps/wallet-dashboard/components/Dialogs/migration/enums/view.enums.ts
@@ -1,0 +1,7 @@
+// Copyright (c) 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+export enum MigrationDialogView {
+    Confirmation = 'Confirmation',
+    TransactionDetails = 'TransactionDetails',
+}
