+++ apps/wallet-dashboard/components/Dialogs/migration/index.ts
@@ -1,0 +1,6 @@
+// Copyright (c) 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './MigrationDialog';
+
+export * from './views';
